# apps/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Owostack API",
    "description": "Unified billing, entitlements, and usage metering API. Owostack sits between your app and payment providers (Paystack, Stripe, Dodo, Polar) to handle subscriptions, feature gating, metered usage, credit systems, and invoicing.",
    "version": "1.0.0",
    "contact": {
      "name": "Owostack",
      "url": "https://owostack.com"
    }
  },
  "servers": [
    {
      "url": "https://api.owostack.com",
      "description": "Production"
    },
    {
      "url": "https://sandbox.owostack.com",
      "description": "Sandbox"
    }
  ],
  "security": [{ "bearerAuth": [] }],
...
  "paths": {
...
    "/v1/customers": {
      "post": {
        "operationId": "createCustomer",
        "tags": ["Customers"],
        "summary": "Create or resolve a customer",
        "description": "Create a new customer or return an existing one matched by email or ID. Updates metadata if the customer already exists.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email"],
                "properties": {
                  "id": { "type": "string", "description": "Optional custom ID" },
                  "email": { "type": "string", "format": "email" },
                  "name": { "type": "string" },
                  "metadata": { "type": "object", "additionalProperties": true }
                }
              },
              "example": {
                "email": "user@example.com",
                "name": "Jane Doe",
                "metadata": { "plan": "enterprise" }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Customer",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Customer" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/v1/customers/{id}": {
      "get": {
        "operationId": "getCustomer",
        "tags": ["Customers"],
        "summary": "Get a customer by ID",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Customer",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Customer" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/v1/entities": {
      "post": {
        "operationId": "addEntity",
        "tags": ["Entities"],
        "summary": "Add an entity",
        "description": "Add a new entity (seat, team member, etc.) scoped to a feature. Enforces plan limits.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["customer", "feature", "entity"],
                "properties": {
                  "customer": { "type": "string" },
                  "feature": { "type": "string", "description": "Feature slug" },
                  "entity": { "type": "string", "description": "Entity ID (e.g. seat ID)" },
                  "name": { "type": "string" },
                  "email": { "type": "string", "format": "email" },
                  "metadata": { "type": "object", "additionalProperties": true }
                }
              },
              "example": {
                "customer": "user_123",
                "feature": "seats",
                "entity": "seat_456",
                "name": "Bob",
                "email": "bob@acme.com"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Entity added",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "entityId": { "type": "string" },
                    "featureId": { "type": "string" },
                    "count": { "type": "integer" },
                    "limit": { "type": "integer", "nullable": true },
                    "remaining": { "type": "integer", "nullable": true }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "get": {
        "operationId": "listEntities",
        "tags": ["Entities"],
        "summary": "List entities",
        "description": "List entities for a customer, optionally filtered by feature.",
        "parameters": [
          { "name": "customer", "in": "query", "required": true, "schema": { "type": "string" } },
          { "name": "feature", "in": "query", "schema": { "type": "string" }, "description": "Filter by feature slug" }
        ],
        "responses": {
          "200": {
            "description": "Entity list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "entities": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": { "type": "string" },
                          "featureId": { "type": "string" },
                          "name": { "type": "string" },
                          "email": { "type": "string" },
                          "metadata": { "type": "object" },
                          "status": { "type": "string", "enum": ["active", "pending_removal"] },
                          "createdAt": { "type": "string", "format": "date-time" }
                        }
                      }
                    },
                    "total": { "type": "integer" }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/v1/entities/remove": {
      "post": {
        "operationId": "removeEntity",
        "tags": ["Entities"],
        "summary": "Remove an entity",
        "description": "Mark an entity for removal. It enters `pending_removal` status and is fully removed at the end of the billing period.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["customer", "feature", "entity"],
                "properties": {
                  "customer": { "type": "string" },
                  "feature": { "type": "string" },
                  "entity": { "type": "string" }
                }
              },
              "example": { "customer": "user_123", "feature": "seats", "entity": "seat_456" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Entity removed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "entityId": { "type": "string" },
                    "count": { "type": "integer" }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "API key in the format `owo_sk_...`"
      }
    },
    "schemas": {
...
      "Customer": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean" },
          "id": { "type": "string" },
          "email": { "type": "string" },
          "name": { "type": "string" },
          "metadata": { "type": "object", "nullable": true },
          "createdAt": { "type": "number" },
          "updatedAt": { "type": "number" }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "enum": [false] },
          "error": { "type": "string" }
        }
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "Authentication failed",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": { "success": false, "error": "Invalid API Key" }
          }
        }
      },
      "BadRequest": {
        "description": "Invalid request",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": { "success": false, "error": "Invalid request body" }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": { "success": false, "error": "Not found" }
          }
        }
      }
    }
  }
}